@@ -11,8 +11,8 @@ namespace Grav\Common\Data;
 use Grav\Common\Grav;
 use Grav\Common\Utils;
 use Symfony\Component\Yaml\Exception\ParseException;
-use Symfony\Component\Yaml\Parser;
 use Symfony\Component\Yaml\Yaml;
+use RocketTheme\Toolbox\Compat\Yaml\Yaml as FallbackYaml;
 
 class Validation
 {
@@ -107,7 +107,7 @@ class Validation
         $method = 'filter' . ucfirst(strtr($type, '-', '_'));
 
         // If this is a YAML field validate/filter as such
-        if ($type != 'yaml' && isset($field['yaml']) && $field['yaml'] === true) {
+        if ($type !== 'yaml' && isset($field['yaml']) && $field['yaml'] === true) {
             $method = 'filterYaml';
         }
 
@@ -128,9 +128,11 @@ class Validation
      */
     public static function typeText($value, array $params, array $field)
     {
-        if (!is_string($value)) {
+        if (!is_string($value) && !is_numeric($value)) {
             return false;
         }
+
+        $value = (string)$value;
 
         if (isset($params['min']) && strlen($value) < $params['min']) {
             return false;
@@ -643,13 +645,21 @@ class Validation
 
     public static function filterYaml($value, $params)
     {
+        if (!is_string($value)) {
+            return $value;
+        }
+
         try {
-            if (is_string($value)) {
-                return (array) Yaml::parse($value);
-            } else {
-                return $value;
-            }
+            return (array) Yaml::parse($value);
         } catch (ParseException $e) {
+            // If YAML compatibility mode is set on, fall back to older YAML parser.
+            if (Grav::instance()['config']->get('system.strict_mode.yaml_compat', true)) {
+                try {
+                    return (array) FallbackYaml::parse($value);
+                } catch (ParseException $e2) {
+                }
+            }
+
             return $value;
         }
     }
